Portals: checkpoint with portal-registry approach
The PortalProvider module gives an portal-definition facade,
returning a React component that can be placed exactly at the site
(destination side) of the portal. When mounted, it registers the
portal's presence with the portal registry.

Portal, in turn, provides an API for projecting content from
anywhere in the page into specifically-named portals. This is
similar to a dependency-injection/IoC mechanism.

 - TopMenuLayout: transformed ContentPortalSlot's to PortalProvider's
 - LayoutExample: added to Examples menu, use Portal convention
 - AboutLayouts: moved to About menu, use Portal convention
 - AboutPlum: use Portal convention, minor text change
 - DocsLayout: use Portal convention

TODO: develop Menu/MenuItem/Submenu further
TODO: probably remove initialize= prop on PortalProvider
TODO: docs