Portal: robust to a race for a mounted portal-registry
Allows {somePortalName} = Portal to be used within the render
function that mounts the <PlumAppRoot>, or even in an outer
component.

If the portal registry is not resolved within 2 seconds, an error
is emitted to the console; also, the existing bad-portal-name rendering
would be shown onscreen at the site of any portal-projected content
that relies on a portal registry never properly created, after diligent
but doomed attempts to resolve 'em.